// src/features/resume/components/contents/components/skills/interfaces.ts
export enum AspectColor {
    GREEN = "green",
    BLUE = "blue",
    RED = "red",
    YELLOW = "yellow",
}

export enum TalentType {
    ENGINEERING = "ENGINEERING",
    STUDY = "STUDY",
    FORM = "FORM",
    THEORY = "THEORY",
    ENTERTAINMENT = "ENTERTAINMENT",
    KNOWLEDGE = "KNOWLEDGE",
    ESSENCE = "ESSENCE",
    APPLICATION = "APPLICATION",
}

export interface AspectColorTypes {
    background: string;
    title: string;
}